Force a six hours interval to retain values
As this plugin does not actually generate values every time that
collectd polls (due to the state file, to avoid duplicates), it makes
sense to artificially increase the retention to a longer time.

In normal conditions (Puppet agent is running periodically) those six
hours should never be overrun as new values will be produced in between.

This will help as well alarming as the "no data" case actually reflects
that something is wrong. If 'puppet_time' has had no fresh data in 6
hours then Puppet has not run at all. OTOH, if 'puppet_time' is
refreshed but there's no new 'puppet_run' this translates into a
compilation failure (that could be used by an alarm, too).