Add ReportAlways configuration parameter
Add a new `ReportAlways` parameter to ensure that Mimir metrics
derived from collectd data are always reported.

The `collectd_puppet_time_ref` metric would stop reporting when
Puppet was disabled, which caused issues with the Alertmanager
alert we had built on top of it. With this configuration option,
we make sure the metric continues to be emitted regardless of
Puppet's state.

We’ve added this as a configurable parameter to enable only in
our hosts. We avoid changing the default behavior for the rest
of the infrastructure, which still relies on the existing
collectd alerting logic.